moved controller class to own folder
